# executor: pre-alloc chunks to optimize `SortExec` in 