--- a/executor/hash_table.go
+++ b/executor/hash_table.go
@@ -198,7 +198,7 @@ func (c *hashRowContainer) GetAllMatchedRows(probeHCtx *hashContext, probeSideRo
 	}
 	var mayMatchedRow chunk.Row
 	for _, ptr := range innerPtrs {
-		mayMatchedRow, c.chkBuf, err = c.rowContainer.GetRowAndAppendToChunk(ptr, c.chkBuf)
+		mayMatchedRow, c.chkBuf, err = c.rowContainer.GetRowAndAppendToChunkIfInDisk(ptr, c.chkBuf)
 		if err != nil {
 			return nil, err
 		}
@@ -256,7 +256,7 @@ func (c *hashRowContainer) GetMatchedRowsAndPtrs(probeKey uint64, probeRow chunk
 	c.chkBufSizeForOneProbe = 0
 
 	for i, ptr := range innerPtrs {
-		matchedRow, c.chkBuf, err = c.rowContainer.GetRowAndAppendToChunk(ptr, c.chkBuf)
+		matchedRow, c.chkBuf, err = c.rowContainer.GetRowAndAppendToChunkIfInDisk(ptr, c.chkBuf)
 		if err != nil {
 			return nil, nil, err
 		}
@@ -299,7 +299,7 @@ func (c *hashRowContainer) GetNullBucketRows(probeHCtx *hashContext, probeSideRo
 	)
 	matched = matched[:0]
 	for _, nullEntry := range c.hashNANullBucket.entries {
-		mayMatchedRow, c.chkBuf, err = c.rowContainer.GetRowAndAppendToChunk(nullEntry.ptr, c.chkBuf)
+		mayMatchedRow, c.chkBuf, err = c.rowContainer.GetRowAndAppendToChunkIfInDisk(nullEntry.ptr, c.chkBuf)
 		if err != nil {
 			return nil, err
 		}