@YangKeao 2023-08-29
Benchmark with `-benchmem`:

```
#### Before Optimize

BenchmarkSortExec
BenchmarkSortExec/(rows:300000,_orderBy:[0_1],_ndvs:_[0_0])-16                 7         151066018 ns/op         2744563 B/op       3322 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0_1],_ndvs:_[1_0])-16                 6         197174333 ns/op         2744793 B/op       3322 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0],_ndvs:_[1_0])-16                  43          23411626 ns/op         2744612 B/op       3323 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[1],_ndvs:_[1_0])-16                   7         151543325 ns/op         2745058 B/op       3324 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0_1],_ndvs:_[10000_0])-16             7         177350741 ns/op         2744530 B/op       3321 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0],_ndvs:_[10000_0])-16               8         131678407 ns/op         2744892 B/op       3324 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[1],_ndvs:_[10000_0])-16               7         150767816 ns/op         2744536 B/op       3321 allocs/op

BenchmarkSortExecSpillToDisk
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0_1],_ndvs:_[0_0])-16                      1        2582932329 ns/op        5193050816 B/op  6307076 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0_1],_ndvs:_[1_0])-16                      1        2647734085 ns/op        5193087784 B/op  6307077 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0],_ndvs:_[1_0])-16                        1        2339302304 ns/op        5193101864 B/op  6307010 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[1],_ndvs:_[1_0])-16                        1        2657485765 ns/op        5193094816 B/op  6307191 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0_1],_ndvs:_[10000_0])-16                  1        2662686436 ns/op        5193106848 B/op  6307151 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0],_ndvs:_[10000_0])-16                    1        2603333640 ns/op        5193150808 B/op  6307208 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[1],_ndvs:_[10000_0])-16                    1        2648335875 ns/op        5192617128 B/op  6306386 allocs/op

#### After Optimize

BenchmarkSortExec
BenchmarkSortExec/(rows:300000,_orderBy:[0_1],_ndvs:_[0_0])-16                 8         138130126 ns/op         2676770 B/op       3295 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0_1],_ndvs:_[1_0])-16                 6         188350159 ns/op         2676557 B/op       3293 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0],_ndvs:_[1_0])-16                  62          19033666 ns/op         2676379 B/op       3293 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[1],_ndvs:_[1_0])-16                   8         141311563 ns/op         2676669 B/op       3294 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0_1],_ndvs:_[10000_0])-16             7         169337388 ns/op         2676491 B/op       3294 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[0],_ndvs:_[10000_0])-16               9         125240971 ns/op         2676476 B/op       3293 allocs/op
BenchmarkSortExec/(rows:300000,_orderBy:[1],_ndvs:_[10000_0])-16               8         141199463 ns/op         2676724 B/op       3295 allocs/op

BenchmarkSortExecSpillToDisk
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0_1],_ndvs:_[0_0])-16                      1        1411099811 ns/op        98868400 B/op    3304978 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0_1],_ndvs:_[1_0])-16                      1        1463131637 ns/op        98866936 B/op    3304969 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0],_ndvs:_[1_0])-16                        1        1213146591 ns/op        98893048 B/op    3305202 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[1],_ndvs:_[1_0])-16                        1        1411506001 ns/op        98878216 B/op    3305064 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0_1],_ndvs:_[10000_0])-16                  1        1445966515 ns/op        98882232 B/op    3305164 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[0],_ndvs:_[10000_0])-16                    1        1490026009 ns/op        98886152 B/op    3305168 allocs/op
BenchmarkSortExecSpillToDisk/(rows:300000,_orderBy:[1],_ndvs:_[10000_0])-16                    1        1425714363 ns/op        98872840 B/op    3304977 allocs/op
```