@YangKeao 2023-08-30
Oops, I realized that `GetRowAndAppendToChunk` doesn't always append the row to the chunk. It's quite misleading. I'll modify the name to `GetRowAndAppendToChunkIfInDisk` and add a new function called `GetRowAndAlwaysAppendToChunk` :facepalm: